Move data masking integration tests to own file
